CRUD ProfessorService

+++ src/main/java/com/project/professor/allocation/fabio1/service/ProfessorService.java
@@ -4,6 +4,7 @@ import java.util.List;
 
 import org.springframework.stereotype.Service;
 
+import com.project.professor.allocation.fabio1.entity.Department;
 import com.project.professor.allocation.fabio1.entity.Professor;
 import com.project.professor.allocation.fabio1.repository.ProfessorRepository;
 
@@ -11,17 +12,21 @@ import com.project.professor.allocation.fabio1.repository.ProfessorRepository;
 public class ProfessorService {
 
 	private final ProfessorRepository professorRepository;
+	private final DepartmentService departmentService;
 
 	public ProfessorService(ProfessorRepository professorRepository) {
 		super();
 		this.professorRepository = professorRepository;
+		this.departmentService = departmentService;
 	}
 
 	// CRUD READ all
-	public List<Professor> findAll() {
-
-		List<Professor> professor = professorRepository.findAll();
-		return professor;
+	public List<Professor> findAll(String name) {
+		if (name == null) {
+			return professorRepository.findAll();
+		} else {
+			return professorRepository.findByNameContainingIgnoreCase(name);
+		}
 	}
 
 	// CRUD READ by ID
@@ -32,10 +37,13 @@ public class ProfessorService {
 		// return allocation;
 		return professorRepository.findById(id).orElse(null);
 	}
+	
+	public List<Professor> findByDepartment(Long departmentId) {
+		return professorRepository.findByDepartmentId(departmentId);
+	}
 
 	// CRUD CREATE
 	public Professor create(Professor professor) {
-
 		professor.setId(null);
 		Professor professorNew = professorRepository.save(professor);
 		return professorNew;
@@ -56,7 +64,7 @@ public class ProfessorService {
 
 	// CRUD DELET by ID
 	public void deleteById(Long id) {
-		if (professorRepository.existsById(id)) {
+		if (id != null && professorRepository.existsById(id)) {
 			professorRepository.deleteById(id);
 		}
